LAN78xx driver: patch for urb interval
This patch was suggested for issues related to
USB polling. Original message:

From: Joshua Emele jemele@acm.org
Date: Wed, 7 Nov 2018 16:07:40 -0800
[PATCH 298/773] lan78xx: Debounce link events to minimize poll storm
The bInterval is set to 4 (i.e. 8 microframes => 1ms) and the only bit
that the driver pays attention to is “link was reset”. If there’s a
flapping status bit in that endpoint data, (such as if PHY negotiation
needs a few tries to get a stable link) then polling at a slower rate
git add drivers/net/usb/lan78xx.c