sc16is uart: suppress rs485 error message
The message seems to be valid only when the SER_RS485_RX_DURING_TX
is not set. In such case the RTS indicates the half duplex data
traffic. If the SER_RS485_RX_DURING_TX is set then the full duplex
is employed and RTS is not used, therefore the error message is
suppressed.