sc16is7xx: Fix for 'Unexpected interrupt: 8'
This fix is ported from here:
https://github.com/raspberrypi/linux/commit/a0514733c3699a98f25b84f69f4d8f9a44c74cc8
Issue details are here:
https://github.com/raspberrypi/linux/issues/2529

Original commit message:
The SC16IS752 has an Enhanced Feature Register which is aliased at the
same address as the Interrupt Identification Register; accessing it
requires that a magic value is written to the Line Configuration
Register. If an interrupt is raised while the EFR is mapped in then
the ISR won't be able to access the IIR, leading to the 'Unexpected
interrupt' error messages.

Avoid the problem by claiming a mutex around accesses to the EFR
register, also claiming the mutex in the interrupt handler work
item (this is equivalent to disabling interrupts to interlock against
a non-threaded interrupt handler).